updated relative to CIF and renamed testing package
